# improve type stability and runtime performance