This is part of an effort to improve the performance of some of our numerical simulation methods for partial differential equations in [Trixi.jl](https://github.com/trixi-framework/Trixi.jl). The changes in this PR improve the performance of our methods by ca. 25%. Thus, it would be really great if this could be merged and a new version of StructArrays.jl could be registered.