@piever 2021-08-25
Thanks for the PR! Did you happen to measure if the compilation overhead when doing `using StructArrays` after this change compared with master?