@ranocha 2021-08-25
Running `julia -e '@time using StructArrays'` with Julia 1.7.0-beta3.0 (2021-07-07) several times yields the following results.
- Current `master`: A typical result is `0.533961 seconds (2.59 M allocations: 182.671 MiB, 3.20% gc time, 0.88% compilation time)`; the times vary slightly by ca. 0.02 seconds
- This PR: A typical result is `0.536773 seconds (2.59 M allocations: 182.672 MiB, 3.23% gc time, 0.88% compilation time)`; the times vary slightly by ca. 0.02 seconds

Thus, the latency of `using StructArrays` did not really change.